getThreshold {InfoFarmSparkML}	R Documentation

getThreshold

Description

get optimal threshold for given predictions

Usage

getThreshold(data, responseCol = "response", predCol = "rate", n = 100,
  logging = FALSE)

Arguments

data

responseColumn
column containing the response variable

predictionColumn
column containing the predictions (double)

n=100
number of thresholds to investigate

logging=FALSE

Format

 NULL

Value

data.frame containing true positives, false positives, true negatives, and false negatives for each investigated threshold and the optimal threshold

[Package InfoFarmSparkML version 1.0 Index]